ModbusSettings: Increase max frame size.
Increase the maximum frame size to 252.

Max words is defined as 125, the modbus spec maximum for "Read Holding Registers"
and "Read Input Registers" requests. These requests require a frame size of
2 + 2 * words (252).

"Write Multiple Registers" requests are limited to 123 words
and require a frame size of 6 + 2 * words (252).

"Read/Write Multiple Registers" requests are limited to 121 words
and require a frame size of 10 + 2 * words (252).